// client/app/src/featchers/users/utils/interfaces/interface.ts
export interface User {
    id: number;
    firstName: string;
    lastName: string;
    image?: string;
    email: string;
    password: string;
    latitude: number;
    longitude: number;
    time: string;
    title: string;
    city: string;
    street: string;
    age: number;
    distance?:number
}

 export interface Location {
    latitude: number;
    longitude: number;
  }
  